refactor: #208 에디터 컨테이너에 불필요한 로직 유틸로 분리
에디터 컨테이너에서 가지고 있지 않아도 될 로직을 다른 파일로 분리했습니다.
